🏗️  structure: reorganize directories by display, action structure #60
